Update: Dilithium tests are implemented on nexys video with the cycle counting and UART/JTAG.
