--- pqc_tests_fpga/cv32a6_ima_sv32_fpga/dilitihum_all/Makefile
@@ -1,31 +1,0 @@
-CC = riscv32-unknown-elf-gcc
-OBJCOPY = riscv32-unknown-elf-objcopy
-
-ARCH = -march=rv32ima_zicsr -mabi=ilp32 -mcmodel=medany
-
-CFLAGS = $(ARCH) -O2 -Wall --specs=nosys.specs
-LDFLAGS = -T ../link.ld -nostartfiles
-
-SRCS = ../../../verif/bsp/crt0.S \
-       ./main.c \
-       ../../../verif/tests/custom/kyber/cbd.c \
-       ../../../verif/tests/custom/kyber/fips202.c \
-       ../../../verif/tests/custom/kyber/indcpa.c \
-       ../../../verif/tests/custom/kyber/kem.c \
-       ../../../verif/tests/custom/kyber/ntt.c \
-       ../../../verif/tests/custom/kyber/poly.c \
-       ../../../verif/tests/custom/kyber/polyvec.c \
-       ../../../verif/tests/custom/kyber/randombytes.c \
-       ../../../verif/tests/custom/kyber/reduce.c \
-       ../../../verif/tests/custom/kyber/symmetric-shake.c \
-       ../../../verif/tests/custom/kyber/verify.c
-
-TARGET = load.elf
-
-all: $(TARGET)
-
-$(TARGET): $(SRCS)
-	$(CC) $(CFLAGS) $(LDFLAGS) -o $@ $^
-
-clean:
-	rm -f $(TARGET)
